1) compact the UI activity and long text with shorthand codes as the following: UIF: update interval frequency THT: Threshold Time THD: Threshold Distance LA: Library Active M: Monitoring C: Connected or not TC: Theta Circ TT: Theta Time TI: Internal Tick TB: Broadcast Tick 2) added more internal states to UI: threshold time, threshold distance, theta time and theta circ. 3) added default value to threshold time and threshold distance. 4) added callbacks to LocationListenerGClient to deliver theta time and theta circ 5)  implements theta time and theta circ callbacks in the activity and update UI. 6) fixed tests and added callbacks. 7) run tests on emulator. 8) integration test on the emulator

+++ gclientlib/src/main/java/com/kimbrough/gclientlib/APIMethods.java
@@ -58,14 +58,14 @@ interface APIMethods {
     /**
      * gets radius of threshold
      */
-    int getThresholdRadius();
+    double getThresholdRadius();
 
     /**
      * sets the radius of threshold distance
      *
      * @param distanceInMeters radius of threshold distance in meters
      */
-    void setThresholdRadius(int distanceInMeters);
+    void setThresholdRadius(double distanceInMeters);
 
     /**
      * gets threshold time
